pyproject.toml
[project]
name = "fragility-monitor"
version = "0.1.0"
description = "AI Market Fragility Monitor"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Fragility Monitor"}]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "requests>=2.31",
  "pyarrow>=14.0",
  "jinja2>=3.1",
  "matplotlib>=3.8",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.5",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "mypy>=1.7",
  "ruff>=0.1",
]

[project.scripts]
fragility = "fragility_monitor.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true

[tool.pytest.ini_options]
pythonpath = ["src"]
